+++ scim2_models/resources/service_provider_config.py
@@ -1,8 +1,6 @@
 from typing import Annotated
 from typing import Any
 
-from pydantic import Field
-
 from ..annotations import Mutability
 from ..annotations import Required
 from ..annotations import Returned
@@ -62,10 +60,7 @@ class AuthenticationScheme(ComplexAttribute):
         httpbasic = "httpbasic"
         httpdigest = "httpdigest"
 
-    type: Annotated[Type | None, Mutability.read_only, Required.true] = Field(
-        None,
-        examples=["oauth", "oauth2", "oauthbearertoken", "httpbasic", "httpdigest"],
-    )
+    type: Annotated[Type | None, Mutability.read_only, Required.true] = None
     """The authentication scheme."""
 
     name: Annotated[str | None, Mutability.read_only, Required.true] = None